refactor(pos): renombrar los componentes PostMain a Post, Post a PostDetail
